+++ CHANGELOG.md
@@ -1,10 +1,24 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.0] - 2024-01-09
+### Added
+- TypeScript support
+- `Paradox.buildElement` now is a module
+    - `Paradox.buildElement` added memoization for html elements
+
+### Fixed
+- `Paradox.pubsub` issue with `unsubscribe` method
+
+### Changed
+- `Paradox.pubsub` cache is now a `Set` instead of an `Array`
+- Internal imports changed from `src` to `build` folder so the new entry point is `build/index.js`
 
 ## [0.2.2] - 2024-01-06
 ### Added